refactor: 💡 Better Auth Plugin
Refactor the better auth plugin and using a new route for examples
